chore: MemberList 개선
